[Module_3][Phase 3: Transcript Calculation][BE_03_039] Transcript Calculation : Group All Block
