+++ src/shared/utils/time.js
@@ -1,0 +1,11 @@
+function TimeNow() {
+  const now = new Date();
+  const hours = String(now.getHours()).padStart(2, "0");
+  const minutes = String(now.getMinutes()).padStart(2, "0");
+  const seconds = String(now.getSeconds()).padStart(2, "0");
+  return `${hours}:${minutes}:${seconds}`;
+}
+
+module.exports = {
+  TimeNow,
+};
